Fix imports for analysis plugins
This means that the base plugins that implement required
functionality are imported and registered explicitly, rather than
implicitly when indigo_za or another locale imports them.
