+++ indigo/analysis/refs/__init__.py
@@ -1,0 +1,1 @@
+import indigo.analysis.refs.base  # noqa
