+++ indigo/analysis/toc/__init__.py
@@ -1,0 +1,1 @@
+import indigo.analysis.toc.base  # noqa
